gtk: Implement gtk::Scrollable for View
This allows View to keep track of the map timestamp itself and keep the
position during scroll speed changes and resizing.

Unfortunately, this change doesn't play nicely with simple scaleY(-1),
so a dedicated downscroll boolean property is introduced instead, along
with buttons for the demo app.

There are two small bugs left.

1. Toggling downscroll while the parent gtk::ScrolledWindow shows
   overshoot restores the adjustment value a frame later so the map
   jumps to the opposite position.
2. Grabbing the scrollbar and starting to drag it makes it jump a bit
   down, which becomes worse as the adjustment value (playfield height)
   increases.

I'm not sure if either of those are my or ScrolledWindow's bugs and I'm
not sure how to work around them in the latter case.