gtk: Make to/from_pixels independent of width
The main issue that this solves is scroll speed depending on the lane
count, however removing the playfield width dependency altogether
results in a valid and simple way to interpret scroll speed, certainly
one among the natural ways to interpret the scroll speed. Notably, this
way shrinking or stretching the playfield past the natural width does
not change the playfield height or the pixels-per-second scroll speed.